+++ Game.Client.Unity/Assets/Scripts/Net/NetworkChannel.cs
@@ -25,6 +25,7 @@ namespace Game.Client
         public event Action<ProgressUpdate> ProgressReceived;
         public event Action<CastInfo> CastReceived;
         public event Action<InventoryUpdate> InventoryReceived;
+        public event Action<WarehouseUpdate> WarehouseReceived;
         public event Action<StatsUpdate> StatsReceived;
         public event Action<BuffUpdate> BuffsReceived;
         public event Action<GoldUpdate> GoldReceived;
@@ -105,6 +106,7 @@ namespace Game.Client
             _connection.On<ProgressUpdate>("Progress", p => ProgressReceived?.Invoke(p));
             _connection.On<CastInfo>("Cast", c => CastReceived?.Invoke(c));
             _connection.On<InventoryUpdate>("Inventory", i => InventoryReceived?.Invoke(i));
+            _connection.On<WarehouseUpdate>("Warehouse", w => WarehouseReceived?.Invoke(w));
             _connection.On<StatsUpdate>("Stats", st => StatsReceived?.Invoke(st));
             _connection.On<LearnedSkills>("Learned", l => LearnedReceived?.Invoke(l));
             _connection.On<SkillBarDto>("SkillBar", b => SkillBarReceived?.Invoke(b));
@@ -284,6 +286,10 @@ namespace Game.Client
 
         public Task SellItemAsync(Guid npcEntityId, Guid instanceId, int quantity) =>
             _connection.SendAsync("SellItem", npcEntityId, instanceId, quantity);
+
+        public Task OpenWarehouseAsync() => _connection.SendAsync("OpenWarehouse");
+        public Task WarehouseDepositAsync(Guid instanceId) => _connection.SendAsync("WarehouseDeposit", instanceId);
+        public Task WarehouseWithdrawAsync(Guid instanceId) => _connection.SendAsync("WarehouseWithdraw", instanceId);
 
         public Task TeleportAsync(Guid npcEntityId, string zoneId) =>
             _connection.SendAsync("Teleport", npcEntityId, zoneId);